Feat path finding for destination
When you make a knight object that has a pos on the board, you can then
see what the shortest path is to a destination position on the board.

The method #build_routes returns a node that has the destination as its
position, and if you can traace back its previous nodes back to the sta-
rting position.